Refactor ntags to have snippets saved in their own file and the original views into the modelviewset by drf
